requirements.txt
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
channels==3.0.3
click==7.1.2
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.3
django-crispy-forms==1.11.2
django-filter==21.1
djangorestframework==3.12.4
gunicorn==20.1.0
h11==0.12.0
httptools==0.1.2
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
iniconfig==1.1.1
Markdown==3.3.4
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.2.0
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
-e git+https://github.com/hotwire-django/turbo-django@b04a3eef185101e6954fd1ae83815d8cb87b4f40#egg=turbo_django
Twisted==20.3.0
txaio==21.2.1
uritemplate==4.1.1
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
webpack-tools==0.2.0
websockets==8.1
whitenoise==5.2.0
zope.interface==5.4.0
